Renamed `GamepadInputProvider` to `HidInputProvider` and rewrote to support sword joystick input
